+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Move session timeout banner to the bottom of the page. Refs STCOR-883.
 * Conditionally use `/users-keycloak/_self` endpoint when `users-keycloak` interface is present. Refs STCOR-835.
 * Wait longer before declaring a rotation request to be stale. Refs STCOR-895.
+* Send the stored central tenant name in the header on logout. Refs STCOR-900.
 
 ## [10.1.1](https://github.com/folio-org/stripes-core/tree/v10.1.1) (2024-03-25)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v10.1.0...v10.1.1)